refactor: reorganise and clean up code
